+++ lib/document_auto_capture/document_auto_capture_result_screen.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 
 import '../disposable_raw_image.dart';
 import '../document_auto_capture/conversion_extensions.dart';
+import '../json_encoder_factory.dart';
 
 class DocumentAutoCaptureResultScreen extends StatelessWidget {
   final DocumentAutoCaptureResult _result;
@@ -17,8 +18,8 @@ class DocumentAutoCaptureResultScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    JsonEncoder encoder = new JsonEncoder.withIndent('  ');
-    String jsonText = encoder.convert(_result.toJson());
+    JsonEncoder jsonEncoder = JsonEncoderFactory.create();
+    String jsonText = jsonEncoder.convert(_result.toJson());
 
     return Scaffold(
       appBar: AppBar(