+++ lib/json_encoder_factory.dart
@@ -1,0 +1,9 @@
+import 'dart:convert';
+
+class JsonEncoderFactory {
+  static const jsonIndent = '  ';
+
+  static JsonEncoder create() {
+    return const JsonEncoder.withIndent(jsonIndent);
+  }
+}