+++ lib/magnifeye_liveness/conversion_extensions.dart
@@ -1,4 +1,5 @@
 import 'package:dot_face_lite/dot_face_lite.dart';
+import 'package:dot_flutter_sdk_example/conversion_extensions.dart';
 
 import '../face/conversion_extensions.dart';
 
@@ -6,18 +7,15 @@ extension MagnifEyeLivenessResultExt on MagnifEyeLivenessResult {
   Map<String, dynamic> toJson() {
     return {
       'bgraRawImage': {
-        'bytes': '${bgraRawImage.bytes.length} bytes',
-        'size': {
-          'width': bgraRawImage.size.width,
-          'height': bgraRawImage.size.height,
-        },
+        'bytes': bgraRawImage.bytes.toJson(),
+        'size': bgraRawImage.size.toJson(),
       },
       'faceDetectorResult': {
         'confidence': faceDetectorResult.confidence,
         'normalizedRectangle': faceDetectorResult.normalizedRectangle?.toJson(),
         'imageParameters': faceDetectorResult.imageParameters?.toJson(),
       },
-      'content': '${content.length} bytes',
+      'content': content.toJson(),
     };
   }
 }