+++ lib/magnifeye_liveness/magnifeye_liveness_screen.dart
@@ -4,12 +4,7 @@ import 'package:flutter/material.dart';
 import '../page_routes.dart';
 import 'magnifeye_liveness_result_screen.dart';
 
-class MagnifEyeLivenessScreen extends StatefulWidget {
-  @override
-  State createState() => _MagnifEyeLivenessScreen();
-}
-
-class _MagnifEyeLivenessScreen extends State<MagnifEyeLivenessScreen> {
+class MagnifEyeLivenessScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -18,17 +13,9 @@ class _MagnifEyeLivenessScreen extends State<MagnifEyeLivenessScreen> {
         ),
         body: MagnifEyeLivenessWidget(
           configuration: MagnifEyeLivenessConfiguration(),
-          onCreated: (controller) {
-            debugPrint(
-                "on created called, starting MagnifEye liveness evaluation");
-            controller.start();
-          },
-          onProcessed: (detection) => debugPrint("on processed called"),
-          onFinished: (result) {
-            Navigator.of(context).pushReplacement(
-                createRoute(MagnifEyeLivenessResultScreen(result)));
-          },
-          onStopped: () => debugPrint("on stopped called"),
+          onCreated: (controller) => controller.start(),
+          onFinished: (result) => Navigator.of(context).pushReplacement(
+              createRoute(MagnifEyeLivenessResultScreen(result))),
         ));
   }
 }