+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+### Added
+- NFC Reading sample.
 
 ## 0.2.0 - 2024-02-16
 ### Added